\name{HW4MC-package}
\alias{HW4MC-package}
\alias{HW4MC}
\docType{package}
\title{
\packageTitle{HW4MC}
}
\description{
\packageDescription{HW4MC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HW4MC}
\packageIndices{HW4MC}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{HW4MC}

Maintainer: \packageMaintainer{HW4MC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
